Define abstract traits and implement Filesystem Concrete Impl.
This commit defines abstractions for Backends that both are both
object-safe and async. In general, a Backend can fetch and persist
state objects, and new up a Journal. A Journal is usually concretely
attached to a particular Backend type i.e. the LocalJournal should
only be used with the LocalBackend.

This commit adds the two above concrete implementations - LocalJournal
and LocalBackend -- though they're not totally complete yet.