Use a Concrete Client instead of a Trait. (#32)
I feel pretty comfortable removing the trait abstraction
for our backend client. Most API libraries use a struct instead
of a trait. And the trait was starting to get annoying as we want
to be able to clone it and have multiple owners. The Arc<dyn trait>
approach was working pretty well, but ultimately I felt like this
was an abstraction we can safely remove without harming the system's
flexability or adaptability in the future.