Implement Observation
This commit uploads observations in parallel. It appears
our API only supports uploading a single observation
at a time, but we collect a possible vector observations,
one for each group, or potentially more depending on
how many were created during the obseravtion period.

This commit sends a request to the backend to upload
a new observation for each observation, but it does
it in parallel for performance.