Generate CLI reference docs automatically (#80)
This commit adds a new executable meant for development
that generates reference documentation for the CLI
in Markdown.

In the future, we'll store the generated docs as an
artifact in CI, and then automatically push it to our
documentation site if we detect a change. For now, we'll
have to manually update the docs when we cut a release.