Introduce multi-core library. (#82)
* WIP: Add multi-core crate to offer utilities for our other services and libraries.

* Introduce mullti-core crate.

There are more and more code shared between the CLI and Aviary.
This commit introduces a new workspace crate, `multi-core`, which
we can use to share code across multiple repos and services.

For initial inclusion, I copied some code that hashes files, which
we use for the CloudFlare API and will use when we're storing
application config.

* Cargo-sort can't handle workspaces.

* Fix missing dependency feature

* Add a tauological test to pass CI