Skeleton outline of a function-based state machine.
This commit introduces the general idea of a function-based
state machine for asking & answering questions to fill in the
manifest. This state machine makes it very easy to isolate
the current context from the rest of the otherwise messy questioning
logic.

Currently, it allows you to pass a mutable Manifest reference around,
letting each state edit the manifest as more information is gathered.
Right now, there is only one implementation; a global implementation
that shows we can define the StateMachine trait on all types.

Next, we plan to narrow this down into function types (closures) that
implement the necessary parameters. This will give us a pleasant
syntactic sugar to define new states as functions.