Add state to check if the user is logged in.
This commit introduces a new state that checks
whether the user is logged in. If they're not, we prompt
them for their login credentials before continuing.
This ensures we have a backend client that we can use to
query the workspaces in this account before continuing with the
init process.