Multi Init: Set the user's workspace.
This commit upgrades the `init` command to handle setting the users
workspace information in the manifest file. The first thing it does,
once we've confirmed the user is logged in, is fetch the workspaces
that the user already has in their account. Then, we list the workspaces
available to them, and ask if they would like to use an existing workspace
or create a new workspace. For now, we don't create the new workspace
for them, but we'll add that soon in a follow-on PR. If they select
one of their existing workspaces, it will be set in their newly
generated `MultiTool.toml` file.