Add (hidden) multi init command. (#111)
This PR adds the new `multi init` command, but hides it from the
CLI help test because it is not yet implemented. You can still run
the command, though the output is not of interest. However, users
won't see this command in the help text, so they can't discover it.
This is kind of like a feature flag for CLI subcommands!