Validate Manifest Missing before Init (#113)
* Validate Manifest Missing before Init

This commit ensures that there is no manifest file found before
initializing a new manifest to prevent us from overwriting an
existing manifest, or accidently dropping the comments in a
manifest.

* Add state to check if the user is logged in.

This commit introduces a new state that checks
whether the user is logged in. If they're not, we prompt
them for their login credentials before continuing.
This ensures we have a backend client that we can use to
query the workspaces in this account before continuing with the
init process.

* Multi Init: Set the user's workspace.

This commit upgrades the `init` command to handle setting the users
workspace information in the manifest file. The first thing it does,
once we've confirmed the user is logged in, is fetch the workspaces
that the user already has in their account. Then, we list the workspaces
available to them, and ask if they would like to use an existing workspace
or create a new workspace. For now, we don't create the new workspace
for them, but we'll add that soon in a follow-on PR. If they select
one of their existing workspaces, it will be set in their newly
generated `MultiTool.toml` file.

* Rename project_dir => application root for consistency.

This commit replaces uses of the word "project" with application
for consistency. It should have no functional change, but the error
messages should now help to guide the user a little bit more
precisely.

* Improve copy in CLI output

This commit is a tiny change that tightens up the copy in one
of our CLI prompts.