Add some helper code to make it easier for Claude to bang out the implementation of `PromptApplication`. (#114)
* Add some helper code to make it easier for Claude to bang out the implementation of `PromptApplication`.

* Multi Init: Implement Workspace Creation

This commit updates the code paths that handle creating workspaces
and applications when the user doesn't want to use an existing
one. We don't totally implement the API for creating an application,
but we get most of the way there.

* Adjust copy in CLI output.

This tiny commit modifies the terminal prompt in `multi init`,
bringing it in line with other similar locales in this codebase.