Server-mode & Controller-mode
Add the ability to run the CLI in Controller mode and Server mode.
Controller mode runs the Kubernetes API controller, which runs the
reconciliation loop to deploy API Gateways (servers). Running in server
mode runs an API Gateway.

Additionally, we watch for changes to Gateway domain objects in addition
to GatewayClass domain objects. We don't currently do anything, but in
the future we'll create an owned deployment that launches the CLI in
Server mode.

This commit also adds a Dockerfile so we can deploy the gateway servers
into the cluster.